+++ cloudbox/config.py
@@ -1,8 +1,33 @@
 import os
+# from flask import Flask
+# from celery import Celery
+import cloudinary
+import boto3
+from botocore.client import Config
 from dotenv import load_dotenv
 load_dotenv()
+
+
+#configure cloudinary
+cloudinary.config( 
+	cloud_name = os.environ.get('CLOUDINARY_CLOUD_NAME'), 
+	api_key = os.environ.get('CLOUDINARY_API_KEY'), 
+	api_secret = os.environ.get('CLOUDINARY_API_SECRET')
+)
+
+#configure boto3
+s3 = boto3.client(
+	"s3", 
+	aws_access_key_id= os.environ.get("AWS_ACCESS_KEY_ID"), 
+	aws_secret_access_key= os.environ.get("AWS_SECRET_ACCESS_KEY"),
+	config=Config(
+		signature_version="s3v4",
+		region_name= os.environ.get("AWS_CLOUDBOX_REGION", "eu-central-1")
+		)
+	)
+
+#configuration variables for the flask app
 class Config:
-    #dbs
     MONGODB_HOST = os.environ.get('CLOUDBOX_NOSQL_DB_URI')
     SQLALCHEMY_DATABASE_URI=os.environ.get('CLOUDBOX_SQL_DB_URI')
     SQLALCHEMY_TRACK_MODIFICATIONS = True
@@ -23,6 +48,30 @@ class Config:
     AWS_SECRET_ACCESS_KEY = os.environ.get("AWS_SECRET_ACCESS_KEY")
     S3_BUCKET_NAME = os.environ.get("S3_BUCKET_NAME")
     S3_BUCKET_BASE_URL = os.environ.get("S3_BUCKET_BASE_URL")
+
+     # CELERY_CONFIG = {
+    #     "broker_url": os.environ.get("CELERY_BROKER_URL"),
+    #     "result_backend": os.environ.get("CELERY_RESULT_BACKEND"),
+    # }
     
 
-    
+
+# def make_celery(app: Flask) -> Celery:
+#     """
+#     Create Celery instance to be used to define tasks
+#     """
+#     celery = Celery(app.import_name)
+#     print(app.import_name)
+#     celery.conf.update(app.config["CELERY_CONFIG"])
+
+#     class ContextTask(celery.Task):
+#         """
+#         Injecting application context to celery's tasks
+#         """
+
+#         def __call__(self, *args, **kwargs):
+#             with app.app_context():
+#                 return self.run(*args, **kwargs)
+
+#     celery.Task = ContextTask
+#     return celery
