+++ cloudbox/store/resources.py
@@ -9,7 +9,7 @@ import os
 from cloudbox.http_status_codes import *
 from cloudbox.models import FolderAsset, FileAsset, BaseAsset
 
-from ..services.upload import upload_file_to_s3, download_file_asset, view_file, download_folder_asset
+from .utils import upload_file_to_s3, download_file_asset, view_file, download_folder_asset
 
 from .fields import (folder_asset_fields, file_asset_fields, asset_viewers_fields, asset_editors_fields,
  asset_general_access_fields)
@@ -137,8 +137,8 @@ class Folder(Resource):
         if restricted_to_owner_editors_general_editors_CUD(asset, user_id):      
             #perform the update if: the user owns the asset, is in the asset's editors list or 
             # the asset's anyone_can_access is editor
-            asset.update(name= args.get('name'))
-            asset.save()
+            asset.modify(name= args.get('name')) #modify updates the asset in place unlike update which
+            #saves the update but requires another query to give the updated documents
             return single_entity_response(asset), HTTP_200_OK
         return NOT_ALLOWED_TO_PERFORM_ACTION_ERROR, HTTP_403_FORBIDDEN
         
@@ -199,9 +199,8 @@ class File(Resource):
         #access given to: verified owners of assets, those in the asset's editors' list, 
         # if the asset's anyone_can_acess is editors
         #id is parent asset id
-        args= file_asset_creation_args.parse_args(strict=True)
+        args= file_asset_creation_args.parse_args()
         user_id= get_jwt_identity()
-
 
         #no id means a user is uploading a file in  their root folder
         parent= FolderAsset.objects.get_or_404(user_id= user_id, parent= None) if id is None else \
@@ -377,8 +376,7 @@ class GeneralAccess(Resource):
         asset= asset= FileAsset.objects.get_or_404(__raw__= {"parent": {"$exists": True}, "_id": ObjectId(id)})
 
         if restricted_to_owner_editors_general_editors_CUD(asset, user_id):      
-            asset.update(anyone_can_access= args.get('access_type'))
-            asset.save()
+            asset.modify(anyone_can_access= args.get('access_type'))
             return marshal(asset, asset_general_access_fields), HTTP_200_OK
         return NOT_ALLOWED_TO_PERFORM_ACTION_ERROR, HTTP_403_FORBIDDEN
